FIX: Use `routeDidChange` event (#9)
`routeWillChange` may fire for transitions which are later aborted (e.g. sidebar links to chat, which then open in the drawer). This can cause split-second flashes of UI, and mess with scrolling. `routeDidChange` does not have that issue, and also means that the UI changes happen in-sync with the rest of the page being refreshed.